final-pjt-back/requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.6
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.6.0
pytz==2021.1
requests==2.28.1
sqlparse==0.4.1
urllib3==1.26.12
